Added RefWorks link option.
Upon being displayed if there is a link to RefWorks it will also be displayed. In order for this to work a new field must be added. To do this: From the admin menu, select struture->content types, under Database select manage fields and add a new field titled Refworks of type text.