Do not list databases that have not been published
If the database is not published, then it will not appear in our A-Z
database list. And if the database is in a database list, but not in the
A-Z list, then that is confusing. So let's err on the side of not
showing users unpublished material (which does not sound like all that
much of an error, after all...)

Signed-off-by: Dan Scott <dscott@laurentian.ca>